+++ esp32/src/main.cpp
@@ -24,6 +24,14 @@ static const int scl_pin = 1;
 // create wifi module
 static ModuleHandler::ModuleHandler mh;
 
+ // create and register the WiFi module
+ static ModuleWiFi wifi;
+
+ // create and register the microSD module
+ static ModuleMicroSD microSD;
+
+static ModuleIrrigation irrigation;
+
 /**
  * @brief Callback for onReceive
  *
@@ -66,21 +74,17 @@ RESET!
 
   Log.noticeln("Starting i2c interface...");
 
-  // create and register the WiFi module
-  // static ModuleWiFi wifi;
-  // mh.RegisterModule(&wifi);
+  mh.RegisterModule(&wifi);
+ 
+  mh.RegisterModule(&microSD);
 
-  // create and register the microSD module
-  // static ModuleMicroSD microSD;
-  // mh.RegisterModule(&microSD);
-
-  static ModuleIrrigation irrigation;
-  mh.RegisterModule(&irrigation);
+  // commented out for now due to conflict with WiFi
+  //mh.RegisterModule(&irrigation);
 
   // start i2c interface
   Wire.onReceive(onReceive);
   Wire.onRequest(onRequest);
-  bool i2c_status = Wire.begin(dev_addr, sda_pin, scl_pin, 100000);
+  bool i2c_status = Wire.begin(dev_addr, sda_pin, scl_pin, 400000);
 
   if (i2c_status) {
     Log.noticeln("Success!");