+++ stm32/Inc/board.h
@@ -14,6 +14,23 @@
 extern "C" {
 #endif
 
+
+/**
+ * @brief Initialize board components
+ *
+ * All peripherials that should be running in normal operation should be
+ * configured here.
+ */
+void Board_Init(void);
+
+/**
+ * @brief De-initialize board components
+ *
+ * All peripherials that should be turned off to save power should be
+ * de-initialized here.
+ */
+void Board_DeInit(void);
+
 /**
   * @brief System Clock Configuration
   * @retval None